+++ lib/ref_inspector/database.ex
@@ -9,11 +9,8 @@ defmodule RefInspector.Database do
   alias RefInspector.Database.Loader
   alias RefInspector.Database.Parser
 
-  @ets_cleanup_delay_default 30_000
-  @ets_data_table_name :ref_inspector_data
-  @ets_data_table_opts [:protected, :ordered_set, read_concurrency: true]
-  @ets_lookup_table_name :ref_inspector_lookup
-  @ets_lookup_table_opts [:named_table, :protected, :set, read_concurrency: true]
+  @ets_table_name :ref_inspector
+  @ets_table_opts [:named_table, :protected, :set, read_concurrency: true]
 
   # GenServer lifecycle
 
@@ -34,25 +31,12 @@ defmodule RefInspector.Database do
   # GenServer callbacks
 
   def handle_cast(:reload, state) do
-    :ok = create_lookup_table()
-    old_ets_tid = fetch_data_table()
-    new_ets_tid = create_data_table()
+    :ok = create_ets_table()
 
-    database_files = Config.database_files()
-    database_path = Config.database_path()
-
-    :ok = do_reload(database_files, database_path, new_ets_tid)
-
-    :ok = schedule_data_cleanup(old_ets_tid)
-    true = :ets.insert(@ets_lookup_table_name, {@ets_data_table_name, new_ets_tid})
-
-    {:noreply, state}
-  end
-
-  def handle_info({:drop_data_table, nil}, state), do: {:noreply, state}
-
-  def handle_info({:drop_data_table, ets_tid}, state) do
-    :ok = drop_data_table(ets_tid)
+    :ok =
+      Config.database_files()
+      |> read_databases()
+      |> update_ets_table()
 
     {:noreply, state}
   end
@@ -64,22 +48,23 @@ defmodule RefInspector.Database do
   """
   @spec list() :: [tuple]
   def list do
-    case fetch_data_table() do
-      nil -> []
-      ets_tid -> :ets.tab2list(ets_tid)
+    case :ets.info(@ets_table_name) do
+      :undefined ->
+        []
+
+      _ ->
+        @ets_table_name
+        |> :ets.lookup(:data)
+        |> Keyword.get(:data, [])
     end
   end
 
   # Internal methods
 
-  defp create_data_table do
-    :ets.new(@ets_data_table_name, @ets_data_table_opts)
-  end
-
-  defp create_lookup_table do
-    case :ets.info(@ets_lookup_table_name) do
+  defp create_ets_table do
+    case :ets.info(@ets_table_name) do
       :undefined ->
-        _ = :ets.new(@ets_lookup_table_name, @ets_lookup_table_opts)
+        _ = :ets.new(@ets_table_name, @ets_table_opts)
         :ok
 
       _ ->
@@ -87,56 +72,35 @@ defmodule RefInspector.Database do
     end
   end
 
-  defp do_reload([], _, _ets_tid) do
+  defp read_databases([]) do
     _ = Logger.warn("Reload error: no database files configured!")
-    :ok
+
+    []
   end
 
-  defp do_reload(files, path, ets_tid) do
-    Enum.each(files, fn file ->
-      database = Path.join([path, file])
+  defp read_databases(files) do
+    database_path = Config.database_path()
 
-      case Loader.load(database) do
-        {:error, reason} ->
-          Logger.info(reason)
+    Enum.map(files, fn file ->
+      entries =
+        database_path
+        |> Path.join(file)
+        |> Loader.load()
+        |> case do
+          {:error, reason} ->
+            _ = Logger.info(reason)
+            %{}
 
-        entries when is_list(entries) ->
-          dataset = {database, Parser.parse(entries)}
-          true = :ets.insert_new(ets_tid, dataset)
-      end
+          entries when is_list(entries) ->
+            Parser.parse(entries)
+        end
+
+      {file, entries}
     end)
   end
 
-  defp drop_data_table(ets_tid) do
-    true =
-      case :ets.info(ets_tid) do
-        :undefined -> true
-        _ -> :ets.delete(ets_tid)
-      end
-
-    :ok
-  end
-
-  defp fetch_data_table do
-    case :ets.info(@ets_lookup_table_name) do
-      :undefined ->
-        nil
-
-      _ ->
-        case :ets.lookup(@ets_lookup_table_name, @ets_data_table_name) do
-          [{@ets_data_table_name, ets_tid}] -> ets_tid
-          _ -> nil
-        end
-    end
-  end
-
-  defp schedule_data_cleanup(ets_tid) do
-    Process.send_after(
-      self(),
-      {:drop_data_table, ets_tid},
-      Config.get(:ets_cleanup_delay, @ets_cleanup_delay_default)
-    )
-
+  defp update_ets_table(datasets) do
+    true = :ets.insert(@ets_table_name, {:data, datasets})
     :ok
   end
 end