+++ lib/mix/tasks/ref_inspector/download.ex
@@ -37,6 +37,7 @@ defmodule Mix.Tasks.RefInspector.Download do
   ]
 
   def run(args) do
+    :ok = start_app()
     :ok = Config.init_env()
 
     {opts, _argv, _errors} = OptionParser.parse(args, @cli_options)
@@ -82,5 +83,15 @@ defmodule Mix.Tasks.RefInspector.Download do
       true -> true
       _ -> Mix.shell().yes?("Download databases?")
     end
+  end
+
+  defp start_app do
+    project = Mix.Project.config()
+
+    if project[:app] not in Application.started_applications() do
+      _ = Mix.Task.run("app.start")
+    end
+
+    :ok
   end
 end